requirements.txt
# Python 3.11 compatible

# Core ML
numpy==1.26.4
ml-dtypes==0.2.0
protobuf==3.20.3          # <-- AGREGAR ESTA LÍNEA
tensorflow==2.15.0
tf-keras==2.15.0

# Computer Vision
opencv-python==4.8.1.78
mediapipe==0.10.8
deepface==0.0.79

# Security
bcrypt==4.1.2

# Image processing
Pillow==10.2.0

# DeepFace dependencies
requests==2.31.0
pandas==2.1.4
gdown==4.7.1
tqdm==4.66.1
mtcnn==0.1.1
retina-face==0.0.14
fire==0.5.0
lightphe==0.0.15